Started api controllers for user.

+++ src/main/java/ru/forum/model/Response.java
@@ -4,15 +4,15 @@ package ru.forum.model;
 @SuppressWarnings("unused")
 public class Response<T> {
 
-    private short code;
+    private int code;
     private T response;
 
-    public Response(short code, T response) {
+    public Response(int code, T response) {
         this.code = code;
         this.response = response;
     }
 
-    public short getCode() {
+    public int getCode() {
         return code;
     }
